chore(e2e): _LCP_INIT_JS + measure_one/page 중복 58줄 → e2e/_perf_helpers.py 추출
두 파일(e2e/test_performance.py, scripts/perf_measure.py)에 동일하게 존재하던
_LCP_INIT_JS / _measure_one() / _measure_page() 를 e2e/_perf_helpers.py 로 추출.
기능 변경 없음.
